cut server
