+++ packages/fullstack/src/lib.rs
@@ -6,44 +6,18 @@
 
 pub use once_cell;
 
-mod html_storage;
-
-pub mod document;
+mod document;
 mod hooks;
-
-#[cfg(feature = "axum")]
-#[cfg_attr(docsrs, doc(cfg(feature = "axum")))]
-pub mod server;
-
-#[cfg(feature = "server")]
-mod render;
-
-#[cfg(feature = "server")]
-mod template;
-
-#[cfg(feature = "server")]
-mod streaming;
-
-#[cfg(feature = "server")]
-mod serve_config;
-
-#[cfg(feature = "server")]
-mod stream;
-
-#[cfg(feature = "server")]
-pub use serve_config::*;
-
-#[cfg(feature = "server")]
-mod server_context;
+mod html_storage;
 
 /// A prelude of commonly used items in dioxus-fullstack.
 pub mod prelude {
     use crate::hooks;
     pub use hooks::{server_cached::use_server_cached, server_future::use_server_future};
 
-    #[cfg(feature = "axum")]
-    #[cfg_attr(docsrs, doc(cfg(feature = "axum")))]
-    pub use crate::server::*;
+    // #[cfg(feature = "axum")]
+    // #[cfg_attr(docsrs, doc(cfg(feature = "axum")))]
+    // pub use crate::server::*;
 
     // #[cfg(feature = "server")]
     // #[cfg_attr(docsrs, doc(cfg(feature = "server")))]
@@ -57,16 +31,12 @@ pub mod prelude {
     // #[cfg_attr(docsrs, doc(cfg(all(feature = "server", feature = "axum"))))]
     // pub use crate::server_context::Axum;
 
-    #[cfg(feature = "server")]
-    #[cfg_attr(docsrs, doc(cfg(feature = "server")))]
-    pub use crate::server_context::{
-        extract, server_context, DioxusServerContext, FromContext, FromServerContext,
-        ProvideServerContext,
-    };
-
-    #[cfg(feature = "server")]
-    #[cfg_attr(docsrs, doc(cfg(feature = "server")))]
-    pub use dioxus_isrg::{IncrementalRenderer, IsrConfig};
+    // #[cfg(feature = "server")]
+    // #[cfg_attr(docsrs, doc(cfg(feature = "server")))]
+    // pub use crate::server_context::{
+    //     extract, server_context, DioxusServerContext, FromContext, FromServerContext,
+    //     ProvideServerContext,
+    // };
 
     pub use dioxus_server_macro::*;
     pub use server_fn::{self, ServerFn as _, ServerFnError};
