--- packages/isrg/Cargo.toml
@@ -1,18 +1,0 @@
-[package]
-name = "dioxus-isrg"
-edition = "2021"
-version.workspace = true
-authors = ["Jonathan Kelley", "Evan Almloff"]
-license = "MIT OR Apache-2.0"
-repository = "https://github.com/DioxusLabs/dioxus/"
-homepage = "https://dioxuslabs.com"
-description = "CLI Configuration for dioxus-cli"
-keywords = ["dom", "ui", "gui", "react"]
-
-[dependencies]
-chrono = { workspace = true }
-http = { workspace = true }
-thiserror = { workspace = true }
-rustc-hash = { workspace = true }
-lru  = { workspace = true }
-tracing = { workspace = true }
