+++ packages/server/src/stream.rs
